requirements.txt
asgiref==3.4.1
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-extensions==3.1.3
django-imagefit==0.6.2
django-mathfilters==1.0.0
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
